chore(cleanup): add custom themes and clean directories
