+++ packages/monogram-ui/.storybook/theme.ts
@@ -1,0 +1,9 @@
+import { create } from "@storybook/theming/create";
+
+export default create({
+  base: "dark",
+  brandTitle: "Monogram",
+  brandUrl: "https://monogram.io/",
+  brandImage: "",
+  brandTarget: "_self",
+});
