+++ packages/monogram-ui/package.json
@@ -16,9 +16,11 @@
     "@storybook/addon-interactions": "^7.1.1",
     "@storybook/addon-links": "^7.1.1",
     "@storybook/blocks": "^7.1.1",
+    "@storybook/manager-api": "^7.1.1",
     "@storybook/react": "^7.1.1",
     "@storybook/react-vite": "^7.1.1",
     "@storybook/testing-library": "^0.0.14-next.2",
+    "@storybook/theming": "^7.1.1",
     "@tailwindcss/typography": "^0.5.9",
     "@types/react": "^18.2.14",
     "autoprefixer": "^10.4.14",
