+++ packages/yolo-ui/.storybook/manager.ts
@@ -1,0 +1,6 @@
+import { addons } from "@storybook/manager-api";
+import theme from "./theme";
+
+addons.setConfig({
+  theme: theme,
+});
