+++ packages/yolo-ui/.storybook/theme.ts
@@ -1,0 +1,9 @@
+import { create } from "@storybook/theming/create";
+
+export default create({
+  base: "dark",
+  brandTitle: "Yolo",
+  brandUrl: "https://resume-web-wheat.vercel.app/",
+  brandImage: "",
+  brandTarget: "_self",
+});
