+++ packages/yolo-ui/package.json
@@ -17,9 +17,11 @@
     "@storybook/addon-links": "^7.1.1",
     "@storybook/addon-onboarding": "^1.0.8",
     "@storybook/blocks": "^7.1.1",
+    "@storybook/manager-api": "^7.1.1",
     "@storybook/react": "^7.1.1",
     "@storybook/react-vite": "^7.1.1",
     "@storybook/testing-library": "^0.2.0",
+    "@storybook/theming": "^7.1.1",
     "@tailwindcss/typography": "^0.5.9",
     "@types/react": "^18.2.14",
     "autoprefixer": "^10.4.14",
